chore: Move tag filter subquery into transactionHelper
