+++ lib/database/models/index.js
@@ -14,7 +14,7 @@
 const EpnRoleSession = require('./epnrolesession');
 const FlpRole = require('./flprole');
 const Log = require('./log');
-const LogTags = require('./logtag');
+const LogTags = require('./logtags');
 const Run = require('./run');
 const Subsystem = require('./subsystem');
 const Tag = require('./tag');
