+++ lib/usecases/log/GetAllLogsUseCase.js
@@ -38,7 +38,7 @@ class GetAllLogsUseCase {
         const { filter, page = {}, sort = { id: 'desc' } } = query;
 
         if (filter) {
-            const { origin, parentLog, rootLog, tag } = filter;
+            const { origin, parentLog, rootLog } = filter;
             if (origin) {
                 queryBuilder.where('origin').is(origin);
             }
@@ -49,29 +49,6 @@ class GetAllLogsUseCase {
 
             if (rootLog) {
                 queryBuilder.where('rootLogId').is(rootLog);
-            }
-
-            if (tag) {
-                const subQueryBuilder = new QueryBuilder();
-                subQueryBuilder.where('tagId').oneOf(...tag.values);
-                subQueryBuilder.orderBy('logId', 'asc');
-                let logTags;
-
-                switch (tag.operation) {
-                    case 'and':
-                        logTags = await await TransactionHelper
-                            .provide(() => LogTagsRepository.findAllAndGroup(subQueryBuilder));
-                        logTags = logTags
-                            .filter((logTag) => tag.values.every((tagId) => logTag.tagIds.includes(tagId)));
-                        break;
-                    case 'or':
-                        logTags = await await TransactionHelper
-                            .provide(() => LogTagsRepository.findAll(subQueryBuilder));
-                        break;
-                }
-
-                const logIds = logTags.map((logTag) => logTag.logId);
-                queryBuilder.where('id').oneOf(...logIds);
             }
         }
 
@@ -86,7 +63,31 @@ class GetAllLogsUseCase {
         const {
             count,
             entities,
-        } = await await TransactionHelper.provide(() => LogRepository.findAndCountAll(queryBuilder));
+        } = await await TransactionHelper.provide(async () => {
+            if (filter && filter.tag) {
+                const subQueryBuilder = new QueryBuilder();
+                subQueryBuilder.where('tagId').oneOf(...filter.tag.values).orderBy('logId', 'asc');
+                let logTags;
+
+                switch (filter.tag.operation) {
+                    case 'and':
+                        logTags = await LogTagsRepository
+                            .findAllAndGroup(subQueryBuilder);
+                        logTags = logTags
+                            .filter((logTag) => filter.tag.values.every((tagId) => logTag.tagIds.includes(tagId)));
+                        break;
+                    case 'or':
+                        logTags = await LogTagsRepository
+                            .findAll(subQueryBuilder);
+                        break;
+                }
+
+                const logIds = logTags.map((logTag) => logTag.logId);
+                queryBuilder.where('id').oneOf(...logIds);
+            }
+
+            return LogRepository.findAndCountAll(queryBuilder);
+        });
 
         return {
             count,
